// src/store/interface.ts
export interface Character {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface ListState {
  results: ReadonlyArray<Character>;
  info: {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
  };
  searchQuery: string | null;
  selectedCharacters: Array<Character>;
  errorMessage: string;
}